// src/types/weapons.ts
export type WeaponType =
  | 'Pistol'
  | 'Pistol1'
  | 'Pistol2'
  | 'Pistol3'
  | 'Blaster1'
  | 'Blaster2'
  | 'Blaster3'
  | 'Revolver1'
  | 'Revolver2'
  | 'Revolver3'
  | 'Rifle'
  | 'Rifle1'
  | 'Rifle2'
  | 'Rifle3'
  | 'GaussGun1'
  | 'GaussGun2'
  | 'GaussGun3'
  | 'SubmachineGun'
  | 'SubmachineGun1'
  | 'SubmachineGun2'
  | 'MachineGun3'
  | 'Machine1'
  | 'Machine2'
  | 'Machine3'
  | 'Shuriken1'
  | 'Shuriken2'
  | 'Shuriken3'
  | 'RocketGun3'
  | 'PlasmoGun2'
  | 'PlasmoGun3'
  | 'Hand'